+++ workers/javascript/packages/plugins/databricks/src/utils.ts
@@ -1,6 +1,5 @@
 import { ConnectionOptions } from '@databricks/sql/dist/contracts/IDBSQLClient';
-import { DatabricksDatasourceConfiguration } from '@superblocks/shared';
-
+import { DatabricksDatasourceConfiguration, ErrorCode, IntegrationError } from '@superblocks/shared';
 import { DatabricksPluginV1 } from '@superblocksteam/types';
 const DATABRICKS_PARTNER_CLIENT_ID = 'Superblocks';
 
@@ -8,6 +7,7 @@ export function getConnectionOptionsFromDatasourceConfiguration(
   datasourceConfiguration: DatabricksDatasourceConfiguration
 ): ConnectionOptions {
   let connectionOptions: ConnectionOptions;
+  
   switch (datasourceConfiguration.connection?.connectionType) {
     case DatabricksPluginV1.Plugin_ConnectionType.M2M:
       connectionOptions = {
@@ -18,6 +18,18 @@ export function getConnectionOptionsFromDatasourceConfiguration(
         oauthClientSecret: datasourceConfiguration.connection?.oauthClientSecret
       };
       break;
+    case DatabricksPluginV1.Plugin_ConnectionType.OAUTH_EXCHANGE: {
+      if (!datasourceConfiguration.authConfig?.authToken) {
+        throw new IntegrationError('OAuth Token Exchange token expected but not present', ErrorCode.INTEGRATION_MISSING_REQUIRED_FIELD);
+      }
+      
+      connectionOptions = {
+        host: datasourceConfiguration.connection?.hostUrl as string,
+        path: datasourceConfiguration.connection?.path as string,
+        token: datasourceConfiguration.authConfig.authToken
+      };
+      break;
+    }
     case DatabricksPluginV1.Plugin_ConnectionType.PAT:
     default:
       // for backwards compatibility, if we do not get a connectionType, we use ConnectionType.PAT