+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## vNext
 - Added support for header value prefix in auth injection flow
+- Added support of Oauth Token Federation auth for Databricks
 
 ## v1.26.0
 - Added support for the `metadata` endpoint in the GraphQL integration